# decklink-output-ui: Pipeline GPU data for preview