@jpark37 2022-09-05
@paulh-aja It looks like the Aja plugin copied the same pattern. You probably want to implement and test similar on your end.